+++ src/pages/season/season_ids.py
@@ -1,0 +1,3 @@
+season_update_label = 'season-update-label'
+dropdown_user_selection_season = 'dropdown-user-selection-season'
+dropdown_user_selection_season_blog = 'dropdown-user-selection-season-blog'